Create the experimental view API
Create the experimental view API
- Add compatibility with vanilla.
- Improve some classes.
- Create 'ApiInit' to centralize initialization logic.
- Implement lookup context.
- Currently experimental, especially the 'Model' interface.

+++ src/main/java/dev/technici4n/fasttransferlib/experimental/api/lookup/ItemLookupContext.java
@@ -4,8 +4,9 @@ import dev.technici4n.fasttransferlib.experimental.api.Content;
 import dev.technici4n.fasttransferlib.experimental.api.transfer.Context;
 import net.minecraft.nbt.CompoundTag;
 
+// TODO set is not a good solution
 public interface ItemLookupContext {
-    int getCount();
+    long getCount();
 
     CompoundTag getTag();
